requirements.txt
--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.5.1/constraints-3.10.txt"

SQLAlchemy==1.4.46
pyodbc==4.0.35
python-dotenv==0.21.0
polars==0.18.3
types-requests==2.28.11.8
types-PyYAML==6.0.12.3
connectorx==0.3.1
apache-airflow-providers-amazon==7.1.0
apache-airflow==2.5.1
pytest==6.2.5
pytest-pythonpath==0.7.4
pytest-cov==4.0.0
isort==5.11.2
black==23.1a1
mypy==0.971
flake8==6.0.0
apache-airflow-client==2.5.1
pre-commit==2.21.0
alembic==1.9.2
psycopg2-binary==2.9.5
cosmic-ray==8.3.7
setuptools== 67.8.0
wheel==0.40.0
pymysql==1.0.3
apache-airflow-providers-http==4.1.1
pyarrow==9.0.0
allure-pytest==2.13.2
apache-airflow-providers-slack==7.2.0
apache-airflow-providers-databricks==4.0.0